requirements.txt
Django==1.7
djangorestframework==2.4.3
drf-extensions==0.2.6
wsgiref==0.1.2
